# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CreateProfileRequest:
    properties:
      address:
        example: 123 Main St, Lagos
        maxLength: 500
        type: string
      avatar_url:
        example: https://example.com/avatars/kaige.png
        maxLength: 2048
        type: string
      bio:
        example: Gopher since 2021. Coffee first.
        maxLength: 1000
        type: string
      phone:
        example: "+15551234567"
        maxLength: 32
        type: string
    type: object
  dto.CreateUserRequest:
    properties:
      age:
        example: 30
        maximum: 100
        minimum: 18
        type: integer
      email:
        example: kaige@example.com
        type: string
      name:
        example: Kaige Saif
        maxLength: 50
        minLength: 3
        type: string
    required:
    - age
    - email
    - name
    type: object
  dto.ErrorEnvelope:
    properties:
      error:
        type: string
    type: object
...
  dto.MessageEnvelope:
    properties:
      message:
        type: string
    type: object
...
  dto.PageMeta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.ProfileEnvelope:
    properties:
      message:
        type: string
      profile:
        $ref: '#/definitions/models.Profile'
    type: object
...
  dto.UpdateProfileRequest:
    properties:
      address:
        example: 123 Main St, Lagos
        maxLength: 500
        type: string
      avatar_url:
        example: https://example.com/avatars/kaige.png
        maxLength: 2048
        type: string
      bio:
        example: Gopher since 2021. Coffee first.
        maxLength: 1000
        type: string
      phone:
        example: "+15551234567"
        maxLength: 32
        type: string
    type: object
  dto.UserEnvelope:
    properties:
      message:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  dto.UsersEnvelope:
    properties:
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.PageMeta'
      user:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  models.Profile:
    properties:
      address:
        example: 123 Main St, Lagos
        type: string
      avatar_url:
        example: https://example.com/avatars/kaige.png
        type: string
      bio:
        example: Gopher since 2021. Coffee first.
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        type: string
      phone:
        example: "+15551234567"
        type: string
      updated_at:
        type: string
      user_id:
        example: 458622d8-daba-4252-8ce1-846277353139
        type: string
    type: object
  models.User:
    properties:
      age:
        example: 30
        type: integer
      created_at:
        type: string
      deleted_at:
        description: |-
          Soft delete: Delete() sets this instead of removing the row.
          Queries filter it out automatically; uniqueness is partial
          (WHERE deleted_at IS NULL) so values stay reusable.
        type: string
      email:
        example: kaige@example.com
        type: string
      email_verified:
        example: true
        type: boolean
      id:
        example: 458622d8-daba-4252-8ce1-846277353139
        type: string
      name:
        example: Kaige Saif
        type: string
      profile:
        $ref: '#/definitions/models.Profile'
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Grip API with JWT auth and email verification.
  title: Grip API
  version: "1.0"
paths:
...
  /users/:
    get:
      parameters:
      - description: Search name and email
        example: kaige
        in: query
        name: search
        type: string
      - default: 1
        description: Page number
        example: 1
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        example: 10
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UsersEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List users with pagination and search
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: User payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - users
  /users/{id}:
    delete:
      parameters:
      - description: User ID
        example: 458622d8-daba-4252-8ce1-846277353139
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      parameters:
      - description: User ID
        example: 458622d8-daba-4252-8ce1-846277353139
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        example: 458622d8-daba-4252-8ce1-846277353139
        in: path
        name: id
        required: true
        type: string
      - description: User payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
...
  /users/{id}/profile:
    delete:
      parameters:
      - description: User ID
        example: 458622d8-daba-4252-8ce1-846277353139
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Delete profile by user ID
      tags:
      - profiles
    get:
      parameters:
      - description: User ID
        example: 458622d8-daba-4252-8ce1-846277353139
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Profile'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get profile by user ID
      tags:
      - profiles
    post:
      consumes:
      - application/json
      parameters:
      - description: User ID
        example: 458622d8-daba-4252-8ce1-846277353139
        in: path
        name: id
        required: true
        type: string
      - description: Profile payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProfileRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProfileEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create profile for user
      tags:
      - profiles
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        example: 458622d8-daba-4252-8ce1-846277353139
        in: path
        name: id
        required: true
        type: string
      - description: Profile payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Profile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Update profile by user ID
      tags:
      - profiles
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and the JWT. Example: "Bearer
      eyJhbGciOiJIUzI1NiJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"