# handlers/user_handler.go
package handlers

import (
	"errors"
	"grip/audit"
	"grip/dto"
	"grip/models"
	services "grip/service"
	"net/http"

	"github.com/gin-gonic/gin"
	"gorm.io/gorm"
)

// CreateUser godoc
// @Summary Create user
// @Tags users
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param request body dto.CreateUserRequest true "User payload"
// @Success 201 {object} dto.UserEnvelope
// @Failure 400 {object} dto.ErrorEnvelope
// @Failure 500 {object} dto.ErrorEnvelope
// @Router /users/ [post]
func CreateUser(c *gin.Context) {
	var req dto.CreateUserRequest

	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": err.Error(),
		})
		return
	}

	user, err := services.CreateUser(req)
	if err != nil {
		internalError(c, err)
		return
	}

	c.JSON(http.StatusCreated, gin.H{
		"message": "User created",
		"user":    user,
	})
	audit.Log(c, models.AuditUserCreate, "users", user.ID, map[string]any{"email": user.Email})
}

// GetUsers godoc
// @Summary List users with pagination and search
// @Tags users
// @Produce json
// @Security BearerAuth
// @Param search query string false "Search name and email" example(kaige)
// @Param page query int false "Page number" default(1) example(1)
// @Param page_size query int false "Page size" default(10) example(10)
// @Success 200 {object} dto.UsersEnvelope
// @Failure 400 {object} dto.ErrorEnvelope
// @Failure 404 {object} dto.MessageEnvelope
// @Failure 500 {object} dto.ErrorEnvelope
// @Router /users/ [get]
func GetUsers(c *gin.Context) {
	var filter dto.UserFilter
	if err := c.ShouldBindQuery(&filter); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": err.Error(),
		})
		return
	}
	filter.Normalize()

	users, total, err := services.ListUsers(filter)
	if err != nil {
		internalError(c, err)
		return
	}

	if len(users) == 0 {
		c.JSON(http.StatusNotFound, gin.H{
			"message": "No users yet",
			"user":    []models.User{},
		})
		return
	}

	c.JSON(http.StatusOK, gin.H{
		"message": "Users found",
		"user":    users,
		"meta":    dto.NewPageMeta(filter.Pagination, total),
	})
}

// GetUser godoc
// @Summary Get user by ID
// @Tags users
// @Produce json
// @Security BearerAuth
// @Param id path string true "User ID" example(458622d8-daba-4252-8ce1-846277353139)
// @Success 200 {object} models.User
// @Failure 404 {object} dto.MessageEnvelope
// @Failure 500 {object} dto.ErrorEnvelope
// @Router /users/{id} [get]
func GetUser(c *gin.Context) {
	id := c.Param("id")

	user, err := services.GetUserByID(id)
	if err != nil {
		if errors.Is(err, gorm.ErrRecordNotFound) {
			c.JSON(http.StatusNotFound, gin.H{
				"message": "User not found",
			})
			return
		}
		internalError(c, err)
		return
	}

	c.JSON(http.StatusOK, user)
}

// UpdateUser godoc
// @Summary Update user
// @Tags users
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path string true "User ID" example(458622d8-daba-4252-8ce1-846277353139)
// @Param request body dto.CreateUserRequest true "User payload"
// @Success 200 {object} models.User
// @Failure 400 {object} dto.ErrorEnvelope
// @Failure 404 {object} dto.MessageEnvelope
// @Failure 500 {object} dto.ErrorEnvelope
// @Router /users/{id} [put]
func UpdateUser(c *gin.Context) {
	var req dto.CreateUserRequest
	id := c.Param("id")

	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": err.Error(),
		})
		return
	}

	user, err := services.UpdateUser(id, req)
	if err != nil {
		if errors.Is(err, gorm.ErrRecordNotFound) {
			c.JSON(http.StatusNotFound, gin.H{
				"message": "User not found",
			})
			return
		}
		internalError(c, err)
		return
	}

	audit.Log(c, models.AuditUserUpdate, "users", user.ID, nil)

	c.JSON(http.StatusOK, user)
}

// DeleteUser godoc
// @Summary Delete user
// @Tags users
// @Produce json
// @Security BearerAuth
// @Param id path string true "User ID" example(458622d8-daba-4252-8ce1-846277353139)
// @Success 200 {object} dto.MessageEnvelope
// @Failure 404 {object} dto.MessageEnvelope
// @Failure 500 {object} dto.ErrorEnvelope
// @Router /users/{id} [delete]
func DeleteUser(c *gin.Context) {
	id := c.Param("id")

	if err := services.DeleteUser(id); err != nil {
		if errors.Is(err, gorm.ErrRecordNotFound) {
			c.JSON(http.StatusNotFound, gin.H{
				"message": "User not found",
			})
			return
		}
		internalError(c, err)
		return
	}

	audit.Log(c, models.AuditUserDelete, "users", id, nil)

	c.JSON(http.StatusOK, gin.H{
		"message":      "User deleted",
		"soft_deleted": true,
	})
}